fix: separate source file into that of core/ROS-node logic
